requirements.txt
Django==2.0
djangorestframework==3.7.4
pandas==0.21.1
pytz==2017.3
requests==2.18.4
